One build issue one test issue.
1) C++ ResultException wasn't compiling without marking "where()"
explicitly as "override" on latest MacOS CLANG.
2) Subtle but fucked-ass issue with any EXT function returning a
PARTIAL success such as TRUNCATED. We cannot use the main error path,
as it could result in an endless cycle should log calls fail, and it has
cyclic dependencies... but we were blindly asserting on anything that
wasn't a success. Added two intelligent assert checks that honor the
CONFIG variables for assertion on ERROR and PARTIAL results.