MAJOR refactor of Gimbal Type system!!
Went back and refactored the entire type system, so that it's more OO
and is not all crammed within the single translation unit that was
gimbal_type.h/.c.

1) gimbal_type.h/.c
  * ONLY shit that's used for type registration along with type
accessors are still in these files. Remodeled to treat GblType more like
an instance in the API.
  * builtin types were removed and are no longer in header and hardcoded
    - GBL_TYPE_XX nomenclature changed to GBL_XX_TYPE which is stored
within its respective header files
  * includes gimbal_builtin_types.h which defines all builtin types and
operations
2) gimal_builtin_types.h/.c
  * contains builtin type INDICES for all builtin types along with
helper macros. This file is the one that should be modified to extend or
modify builtin types.
3) gimbal_value_types.h/.c
  * header file simply defining types for all primitive/value types
(using builtin indices)
4) gimbal_type_.h
  * private header containing all that was previously static state
within gimbal_type.c, so that symbols can be used in other sources.
  * added utility macros for class flags and MetaClass/Type conversions
5) gimbal_class.h/.c
  * contains all GblClass_* instance methods and functionality
  * contains GBL_CLASS_CAST/TRY/CHECK utility macros
  * GblType payload has been changed to an opaque private type because
it's now hidden behind bit magic.
  * went through fucking hell and back again to be able to use 6 bits
from the GblType class pointer by ensuring MetaClass alignment is on
64-byte boundaries.
    - bits are used for GBL_CLASS_FLAGS_
       * moved IN_PLACE/OWNED/FLOATING logic from GblObject to GblClass,
here.
           - still need to finish fixing GblObject to not use it
6) gimbal_instance.h/.c
  * contains all GblInstance* instance methods and functionality
  * conatins GBL_INSTANCE_CAST/TRY/CHECK utility macros
  * need to finish implementing floating classes and sinking
7) gimbal_interface.h
  * lightweight inline wrapper around GblClass with utility macros for
extracting outer classes and doing interface-y shit

***** ALL UNIT TESTS ARE PASSING BUUUUUT *******
a. Floating classes and instance ownership is all kinds of fucked and is
probably just leaking.
b. GblObject is duplicating GblClass functionality now with its flags.