Fixed MinGW crash on property deletion!
This was a fucking horrible one. Unregistering properties for a given type was referencing the implicit property root meta info node after it had already been deleted, potentially looping further or shitting on pointer accesses. Removing the last property automatically removes the root node as well, so the loop had to be revised.

How it never crashed once and worked fine on every other platform is beyond me.