Removed pthreads and C atomics dependencies
1) Threads
* using a minimalist gitlab project called TinyCThreads which is
basically just pthreads on every non-Win32 platform, where it's native
windows threads. Everything is exposed via the C11 API, so in the
future, when Mirosoft extracts its cock from its asshole, it should be
easy to swap it out for real C11 threads.
2) Atomics
Due to the fact that
  1) C11 and C++11 atomics are NOT compatible
  2) Win32 doesn't support C11 atomics
I have had to take the drastic measures of making an abstraction layer
in api/preprocessor/gimbal_atomics.h. For Windows it's just a volatile
uint16 with Inc/Dec Win32 functions. For CPP11, it uses std::atomic<>.
For C11, it uses stdatomic.h types. Since the APIs are inconsistent,
every type has its own set of accessors:
  * GBL_UINT16_INIT
  * GBL_UINT16_LOAD
  * GBL_UINT16_STORE (etc)
As such,
  1) I am only adding what I need which is only uint16_t now
  2) the implementation WILL NOT WORK without being:
    a. Win32
    b. C++11 with <atomics>
    c. C11 with <stdatomic>