Slowly making my way to Windows support
Fixed a whole slew of fucking annoying issues still working on getting Visual Studio to build. It officially builds all the C shit and is now shitting on C++ headers and unit tests.

1) Had to stop #defining max_align_t on Windows, because it interfered with someshit Qt was doing and was a terrible idea anyway. Use GBL_MAX_ALIGN_T now.
2) Added .vs metadata bullshit to gitignore
3) MSVC implementation of gimbal_atomics.h was fucked for multiple reasons
    * it only supports SIGNED int16 atomics, not unsigned
    * InterlockedInc/Dec are prefix operations while std C is postfix
4) Fucking Windows.h includes some other bullshit which literally has enumerated values called ERROR which 20 layers away causes build errors with GBL_API validation macros. #define NOGDI had to be added before any windows.h bullshit inclusion.
5) Moved GblQuark to use tinycthreads

** TINY C THREADS DONT EVEN FUCKING COMPILE WITH MSVC ON WINDOWS **

Had to hack in support for MSVC because they're retarded. Have to remove it as a submodule and own it myself I guess.