Fixed build failure on Linux/Clang
It was totally right, I was not returning anything from
GblContext_callRecordSet_ which was causing an assert to propagate into
tests where it was caught as a failure.

Also got rid of GBL_API_FUNCTION() line of macros, considering MSVC can
handle all of the clusterfuckery around the autogen macros, but it can't
handle that pattern... must be a real piece of shit!