Fixed GblThread on Windows!
Oh. My. GOD. Apparently GetCurrentThread() on Windows returns a psuedo-handle to the current thread which does not allow it to be stored and associated with a particular thread, since it will always reflect THE CURRENT THREAD. The concurrency model was having every thread initialize its own thread handle, then later, the spawning thread was using that handle to join the thread... which caused an infinite loop as it essentialy joined on itself.

Solution? ONLY the main thread uses that psuedo-index, other threads have their handles stored from the second they are created. This does create a corner case ONLY ON WINDOWS where a spawned thread can never refer back to the main thread's handle and operate on it, but like... for now there is no alternative that can still adhere to the C11 API. If this is a problem in the future, need to reevaluate the entire threading abstraction, because Windows does not play nicely with the C11 model