NO MORE PUBLIC INLINE FUNCTIONS, ALL EXPORTED!!
** With the two exceptions of **
1) Some of the GBL_CTX_X() macros withing core/gimbal_ctx.h rely on some
   internally defined inline functions to proxy calls. This is getting
reworked anyway
2) The unit testing macros such as GBL_TEST_VERIFY() and
   GBL_TEST_COMPARE() are fundmentally built upon (the GBL_CTX_X())
macros. Will have to create and export actual top-level entry-points for
gblTestCompareX() for other languages to bind to.

***** THIS GIVES US ****
- Better interoperability with other languages. The statically compiled
  ones like Jai cannot bind to inline functions, because there's no
exported symbol to link to
- Cleaner header files
- Probably smaller code
- Less header dependencies
- Faster compile-times

If this is a problem, I suggest building with LTO enabled. This is what
I'm doing for Sega Dreamcast builds, and the performance gains even with
liberal inlining is huge.

NOTEWORTHY THINGS:
1) gimbal_sort.h never had a corresponding .c file, because it was all
   inlined. Now it does.
2) Same goes for gimbal_nary_tree.h
3) Same goes for gimbal_linked_list.h